Rewrite fmodmusic and fmodsoundfx for fmodex 4.44.41
Music and sound playback now works, though there should probably be
more error checking.

Except for crippling slowdown at the hands of something I haven't
managed to diagnose completely yet, but appears to be related to
copying SDL surfaces around too much, the game is in good working
condition.

Besides that, the only problem I've noticed is that the
player's/opponent's score does not actually appear in the score box.
Other text may be missing as well.

Oh, and the cursor appears as a question mark in the pause menu.