Nuke QuickFade.
QuickFade used SDL_SetGammaRamp by way of SDLU_SetBrightness, which is
bad because it sets the gamma for the entire display, not just the
window the main SDL surface is in. SDLU_SetBrightness has been removed
because it is Considered Harmful.

Also, its (completely unused) color argument was the only instance of
MRGBColor in the entire codebase, so that can get chucked out, too.

The QuickFade function stubs have been left in the code for when
non-bad replacements are written (there is a FadeScreen function in
hiscore.cpp that may be a decent candidate).