Score: fix display.
The score and level failed to display because the mask image for the
score font was saved as a grayscale PNG (8-bit) rather than as an
indexed 1-bit, 2-color PNG that the code expected. This caused the mask
to be loaded incorrectly, in such a way that all of the glyphs were
masked out.

Thanks, GIMP, for fixing what graphicsmagick didn't. I guess you're
only mostly a huge heap of garbage.

Also, this sure was an annoying bug to track down, but at least I
learned a few things.