MTypes: replace MPoint with SDLU_Point.
MPoint has been replaced with SDLU_Point, which conveniently should be
able to be simply replaced by SDL_Point when the upgrade to SDL2 occurs.

The fields of the struct have changed from v and h to y and x, which I'm
pretty sure is more standard or less confusing or something.

As a result of the fields changing and C++ allowing it, the code will be
using struct field names during initialization from now on. I think it's
the better way of doing things anyway.

Hopefully there are no visual or functional changes as the result of
careless find/replacing. Brief testing seems to indicate things are in
order.