SDLU: crazy, bad speedhacks.
By making sure that the blitted surface is only flipped to the screen
at most once every 2 ticks, we can achieve something that is very
nearly playable.

And also running at half the framerate. IMPORTANT NOTICE: this is
absolutely not the correct way to fix the crippling slowdown the game
suffers during normal play. But this works for now, until I can tackle
the (apparently) systemic problems that actually cause the slowdown.