fmod: update API usage for FMOD core
This was a surprisingly simple change. Good on the FMOD team for
keeping their APIs pretty consistent over the course of probably more
than a decade. I'm interested in porting the audio system to something
that's more open source, but this was the only actual change needed to
be able to compile and run this on macOS Monterey.

I'm actually more astonished that SDL 1 still functions, considering
how long it has not been maintained, but maybe the homebrew versions
secretly have patches. I'm too lazy to check, as our first item of
business will be to port this to SDL 2, which is probably going to be
a bumpy ride.