Source: "convert" to C99
This consisted of renaming the files, inserting the stdbool.h include
where appropriate, minor tweaks to the makefile, and converting one
global value initializer (back) to a macro.

In the 8 years since I last set my sights on this codebase, I have
formed the opinion that I don't care for C++ as a programming
language. Thus, there is no reason to pretend that this is C++ code
(it never was: out-of-order designated initializers don't work in C++
because of the complexity of constructors, and they didn't even exist
in the language until C++20, so this was always using a nonstandard
GNU extension anyway).