Source: "port" to SDL2
This is about the bare minimum necessary work to get the game to
function compiled against SDL2. I was actually surprised at how little
work it was, with the stickiest problem being the input handling.
There are some minor graphical glitches due to some rects no longer
properly getting erased, which is interesting, and the window does not
currently properly handle anything that causes the main surface to
update (such as resizing it or moving it between (different DPI?)
monitors).

Using scare quotes on two commits in a row feels bad, but hell, I can't
consider this a true porting job because it seems like it's not really
doing things nicely. It seems like the blitting in particular can
probably be offloaded to SDL and dumped onto the graphics hardware.
This may not be as trivial as it sounds, as it will likely require
cleaning up the resources, which will mean rewriting the code that
handles the resources. Probably not very fun, but likely the main
thing that will make it actually run properly.

Also I've left some debug logging in because I don't really care.