prefs: use SDL to get platform-appropriate prefs file
This still just dumps structs to disk, but it's slightly more clever
about how it does so. I simplified some of the parsing code by
replacing some bit twiddling with memory comparisons. It would be nice
to be able to replace the hexadecimal parsing/dumping, but it's the
only thing that guarantees the prefs key names don't show up in the
values, so writing direct binary data would be unsafe from the
standpoint that it could generate a file that wouldn't parse
correctly.