options: implement the rest of the features
This is pretty much everything I wanted except for proper help text
generation, though the structure is in place for that and the main
thing that's left is the actual work.

It would be reasonable to implement specifying long option values as
`--option=value`, which is a common style. There's also the matter of
allowing multiple values for options/arguments. I haven't done much
scaffolding there, but did produce the start of an idea.

This has certainly been a tour de force of insane metaprogramming, but
I feel like that's one of the great strengths of zig in general. There
were several times where I was pretty sure I had backed myself into a
corner, design-wise, but I was able to figure out a solution (the most
miraculous of which is that the method-style function calls allow
uniform constant value access across both instances and struct
declarations, assuming you don't need both from the same type). The
compile-time duck-typing is a super powerful tool that has certainly
given me enough rope to hang myself. The main thing that impressed me
with this is that the functionality of the language itself is pretty
dang intuitive. The only real snags I hit were a few compiler bugs
(to be expected, I guess, given the current (lack of) maturity of the
language). There was (and still is) some challenge in figuring out how
to express the concepts in my head into the constraints of the
language. It's hard to resist the siren song of C's get out of jail
free void pointer and deal with the constraints of strongly typed
everything (at runtime, at least).

I am a little sad about how copy-pasted some of this code looks (and it
certainly suffers from the arrow antipattern, if you subscribe to
that), but as far as I can figure, it's not really reasonably possible
to extract it out the functionality: it's too entangled into the flow
control and there are deeply buried unique pieces of logic that don't
make sense to try to abstract over. But hey, All things said, not too
bad. It seems that this is a pretty flexible command-line-application
toolkit in just under 600 lines of code, which is pretty good in my
book.

I don't know how I feel about the fact that I did need to switch
everything to a callback style to support the subcommands.
Well, "need" is perhaps a strong word. It was certainly the most
obvious approach, though.